transform_docker/requirements.txt
aiohttp==3.7.4.post0
aiodns==2.0.0
liquidpy==0.6.3